Quantum Harmonic Oscillator
The quantum harmonic oscillator is a fundamental concept in quantum mechanics that describe a particle behaviour when trapped in parabolic potential well. It's similar to a spring mass problem, but with quantized state of energy levels. Each energy level corresponds to a unique wave function, or "state," representing the probability distribution of the particle's position. The ground state, or lowest energy state, has a nonzero energy (called the zero-point energy), reflecting the inherent uncertainty in the particle's position and momentum due to quantum mechanics. This model is essential in describing vibrational modes in molecules, phonons in crystals, and many other quantum systems.