Simulating Quantum Harmonic Oscillator (QHO)
Quantum harmonic oscillator (QHO) is fundamental concept in quantum mechanics (QM) that describe a particle behaviour when trapped in parabolic potential well. It's similar to spring mass problem in classical physics, but the difference is that the energy levels are quantized. meaning the particle can only exist at specific energy levels, which are evenly spaced. Each energy level corresponds to a unique wave function, or "state," representing the probability distribution of the particle's position. The ground state, or lowest energy state, has a nonzero energy (called the zero-point energy), reflecting the inherent uncertainty in the particle's position and momentum due to quantum mechanics. This model is essential in describing vibrational modes in molecules, phonons in crystals, and many other quantum systems.